potentially fix discord presence corruption?
This issue has been going on for a while and I'm not sure where the issue comes from. It is very tedious to debug, though changing the calls order seems to have fixed it. It could potentially be an issue with resources being accessed at the same time by both the main thread and the discord thread, but we'll see.