discord rpc connection is now updated in the main thread
It seems like race conditions were the cause for presence data to occasionaly be corrupted. I didn't feel like spending too much time on making the thread's job safe to avoid these issues, and I also realized the Discord API is fast enough not to cause lag if it is updated in the main loop. So from now on, the Discord RPC connection will be updated in the main thread.